requirements.txt
matplotlib==3.5.1
numpy==1.21.2
opencv_python_headless==4.5.5.62
pandas==1.3.5
Pillow==9.1.1
pydicom==2.3.0
pynvml==11.4.1
scikit_image==0.19.1
scikit_learn==1.1.1
scipy==1.7.3
seaborn==0.11.2
skimage==0.0
torch==1.10.0
torchsummary==1.5.1
torchvision==0.11.0
tqdm==4.61.2
wandb==0.12.6
